Combine notification services
`AppNotificationService` and `EmailNotificationService` have a lot of overlapping
logic: for most notification events, they query the same data to send
notifications to the same set of users. Having to implement the same logic in two
different places makes it more work to add a new notification to the system.

Combine the two classes into a single `NotificationService` that handles both
in-app and email notifications.

The tests remain separate, though the test classes are moved and renamed; the
two test suites are too different to be worth trying to combine.
